각각의 ViewModel마다 사용될 State를 각각 다르게 정의해서 사용하게 구현
Koin 모듈을 정의하는 클래스를 DailyMapApplication -> DropApplication으로 변경
