--- app/src/main/java/com/dev/playground/dailymap/DailyMap.kt
@@ -1,29 +1,0 @@
-package com.dev.playground.dailymap
-
-import android.app.Application
-import com.dev.playground.data.di.localDataSourceModule
-import com.dev.playground.data.di.repositoryModule
-import com.dev.playground.domain.di.dispatcherModule
-import com.dev.playground.domain.di.useCaseModule
-import com.dev.playground.presentation.di.viewModelModule
-import org.koin.android.ext.koin.androidContext
-import org.koin.core.context.startKoin
-
-class DailyMap: Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-
-        startKoin {
-            modules(
-                viewModelModule,
-                dispatcherModule,
-                useCaseModule,
-                repositoryModule,
-                localDataSourceModule,
-            )
-            androidContext(this@DailyMap)
-        }
-    }
-
-}
