--- presentation/src/main/java/com/dev/playground/presentation/util/UiState.kt
@@ -1,11 +1,0 @@
-package com.dev.playground.presentation.util
-
-sealed class UiState<out T> {
-
-    data class Success<out T>(val data: T?) : UiState<T>()
-
-    data class Failure(val exception: Throwable): UiState<Nothing>()
-
-    data class Loading(val loading: Boolean = false): UiState<Nothing>()
-
-}
